Adding Profile table and make single_project page
Now we're ready to assign each projects to website users. Creating Profile table helps us to do that. Project and Review tables need to be connected with this new table. Notive Profile table is inside Users app but those two are inside Projects app.After that I've made a new template so that each project is shown indeviduaslly.