Add some signal to Users app
Changes on User and Profile tables need to be synchroned with each others. I want whenever a new user joins the website a profile is created for him. I also want whenever the user updates his profile his row on User table is updated simoltaneously. Because of one to one relationship which is stablished on Profile table if a User is wiped out from website's database its related Profile will be deleted. At the last signal we want this reversly. Wheneve a Profile is deleted its related User is deleted.