Moving search functionalities into seperate files.
Bunch of code lines in views.py makes it messy. Inserting search functionality into seperat file reduce some of this mess. In "single_project.html" there was a link to the author of project but it wasn't refer to anything. I addressed this link so as now it refers to the author.