Solve pagination and search combining problem
When we combine search and pagination a problem comes up. Only the first page of searched content is shown correctly. When we click on the secon button it shows us the general second page not the second page of result. Using JavaScript codes I added an input tag into search form. This input tag is hiden but because we're using get method the value of this input element will be added into url. This process results in a url which includes both page number and seraching content.