Tweek review functionality
When an Unauthenticate user goes on a project he/she sees that can't vote. Site configuration requires him/her to login in order to vote on projects. It would be nice if as a user logs in to the site he will be redirected to the page which he/she was on before loging. In this case as user logs in, he'll be redirected to the single_project page so as vote on that project he was banned from voting on already.